requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
about-time==4.2.1
    # via alive-progress
alive-progress==3.3.0
    # via -r requirements.in
asgiref==3.9.1
    # via django
django==5.2.4
    # via
    #   -r requirements.in
    #   django-filter
    #   djangorestframework
    #   djangorestframework-simplejwt
django-filter==25.1
    # via -r requirements.in
djangorestframework==3.16.0
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==5.5.1
    # via -r requirements.in
graphemeu==0.7.2
    # via alive-progress
gunicorn==23.0.0
    # via -r requirements.in
numpy==2.3.1
    # via pandas
packaging==25.0
    # via gunicorn
pandas==2.3.1
    # via -r requirements.in
pillow==11.3.0
    # via -r requirements.in
pyjwt==2.10.1
    # via djangorestframework-simplejwt
python-dateutil==2.9.0.post0
    # via pandas
pytz==2025.2
    # via pandas
six==1.17.0
    # via python-dateutil
sqlparse==0.5.3
    # via django
tzdata==2025.2
    # via
    #   django
    #   pandas
